Add item model and update db and services

+++ backend/app/database/database.py
@@ -2,7 +2,10 @@ from sqlmodel import Session, SQLModel, create_engine
 
 from app.config import get_settings
 
-from app.database.models import Item
+from app.database.models import Item as ItemDB
+
+from app.models.item import Item
+
 
 database_url = get_settings().dev_database_url
 
@@ -20,16 +23,34 @@ def get_session():
         yield session
 
 
-def add_item(session: Session, id: str, access_token: str):
-    session.add(Item(id=id, access_token=access_token))
+def add_item(session: Session, item: Item):
+    session.add(
+        ItemDB(
+            id=item.id,
+            institution_name=item.institution_name,
+            access_token=item.access_token,
+        )
+    )
     session.commit()
 
 
 def get_item(session: Session, id: str):
-    return session.get(Item, id)
+    item_db = session.get(Item, id)
+
+    if item_db is None:
+        return None
+
+    return Item(
+        id=item_db.id,
+        access_token=item_db.access_token,
+        institution_name=item_db.institution_name,
+    )
 
 
 def delete_item(session: Session, id: str):
     item = session.get(Item, id)
+    if item is None:
+        return
+
     session.delete(item)
     session.commit()
